Turbolinks removed. It was causing more problems than benefits.
Fix for magnify-large to be always on top (z-index:999)

Rows of products are displaying only three elements rather than using only one div with 'row' class and break the elements inside. With this, it is possible to easily set the columns of the rows with the same height, giving a very beautiful look.