A client will only see products which are available. And 'not found product' page is displayed if he tries to do directly by the url.
A proper client page to show a product. It is not full implemented yet.

Title of products link to their exhibition.

Short description added to product model. Previous product_description column was renamed to product_long_description.

Product model has a method price, which returns the unitary price of the product multiplied by the package quantity.