Rubocop run with -a option.
Added pagination to search results.

Breadcrumbs for each product's categories are added when displaying an individual product. It helps the shopper to have a better notion where that product can be found. Also, it offers the ability to quickly navigate to similar products.

Search results are better formatted on the screen.